Refactor session metrics and errors (#613)
Refactors session metrics to use static variables, this saves us 80
bytes per session (plus some more from being cloned down from proxy).
The one downside is that I had to remove one test, because using static
variables for metrics doesn't work well the with multithreaded tests.
But the test was simple enough that I think it's worth losing that small
bit of coverage.

It also refactors session's errors to be more helpful, and include more
information when logged. I did this through a through a trait so that
later we can make it easier to log other errors and types without
duplication. I also refactored the branches in the packet send code to
ensure that `PACKETS_DROPPED` is always being incremented when we don't
send packets.